# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "车险报价对比助手 API",
    "description": "MVP 后端接口；FastAPI 自动生成的 OpenAPI 是唯一请求/响应契约。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/quotes": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Create Quote",
        "description": "创建报价容器：MANUAL 直接进入待确认；UPLOADED 只建 DRAFT 容器。",
        "operationId": "create_quote_api_projects__project_id__quotes_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "Get Quote",
        "description": "报价完整结构化数据（各层明细、置信度、evidence、冲突信息）。",
        "operationId": "get_quote_api_quotes__quote_id__get",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "quotes"
        ],
        "summary": "Update Quote",
        "description": "编辑基本信息与价格分项；价格变更后事务内重算。",
        "operationId": "update_quote_api_quotes__quote_id__patch",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quotes"
        ],
        "summary": "Delete Quote",
        "description": "删除报价及全部明细（文件资产按无引用规则由 TASK-03 清理）。",
        "operationId": "delete_quote_api_quotes__quote_id__delete",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}/confirm": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Confirm Quote",
        "description": "确认报价：价格分项完整性校验 + 车辆摘要冲突二选一。",
        "operationId": "confirm_quote_api_quotes__quote_id__confirm_post",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}/coverages": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Create Coverage",
        "description": "新增险种行（基础车险/附加险/未识别项统一入口）。",
        "operationId": "create_coverage_api_quotes__quote_id__coverages_post",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}/coverages/{row_id}": {
      "patch": {
        "tags": [
          "quotes"
        ],
        "summary": "Update Coverage",
        "description": "编辑险种行；补标准码即完成未识别项的手动映射。",
        "operationId": "update_coverage_api_quotes__quote_id__coverages__row_id__patch",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quotes"
        ],
        "summary": "Delete Coverage",
        "description": "删除险种行（含未识别项的“丢弃”操作）。",
        "operationId": "delete_coverage_api_quotes__quote_id__coverages__row_id__delete",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}/services": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Create Service",
        "operationId": "create_service_api_quotes__quote_id__services_post",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}/services/{row_id}": {
      "patch": {
        "tags": [
          "quotes"
        ],
        "summary": "Update Service",
        "operationId": "update_service_api_quotes__quote_id__services__row_id__patch",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quotes"
        ],
        "summary": "Delete Service",
        "operationId": "delete_service_api_quotes__quote_id__services__row_id__delete",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}/packages": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Create Package",
        "description": "新增保障包（可携带内部保障列表，单事务创建）。",
        "operationId": "create_package_api_quotes__quote_id__packages_post",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}/packages/{package_id}": {
      "patch": {
        "tags": [
          "quotes"
        ],
        "summary": "Update Package",
        "operationId": "update_package_api_quotes__quote_id__packages__package_id__patch",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Package Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quotes"
        ],
        "summary": "Delete Package",
        "operationId": "delete_package_api_quotes__quote_id__packages__package_id__delete",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuoteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_QuoteRead_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuoteRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiResponse[QuoteRead]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuoteCreate": {
        "properties": {
          "insurerCode": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Insurercode",
            "description": "保险公司标准码"
          },
          "insurerName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Insurername",
            "description": "其他公司的自由输入名称"
          },
          "agentName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentname",
            "description": "保险员称呼（可选）"
          },
          "source": {
            "$ref": "#/components/schemas/QuoteSource",
            "description": "MANUAL 创建即 PENDING_CONFIRM；UPLOADED 只建 DRAFT 容器"
          }
        },
        "type": "object",
        "required": [
          "insurerCode",
          "source"
        ],
        "title": "QuoteCreate",
        "description": "创建报价容器：预置公司存预置码，其他公司固定 OTHER 并必填公司名。"
      },
...
      "QuoteSource": {
        "type": "string",
        "enum": [
          "UPLOADED",
          "MANUAL"
        ],
        "title": "QuoteSource",
        "description": "报价来源：上传解析 / 纯手动录入。"
      },
...
    }
  }
}